@@ -6,7 +6,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.2.1+56
 
 environment:
-  sdk: '>=2.18.0 <3.0.0'
+  sdk: '>=3.0.3 <4.0.0'
 
 dependencies:
   flutter:
@@ -19,7 +19,7 @@ dependencies:
   graphql_flutter: ^5.1.2
   shared_preferences: ^2.2.2
   intl: ^0.18.1 #to format dates
-  pubnub: ^4.3.0
+  pubnub: ^4.3.1
   flutter_dotenv: ^5.1.0
   emoji_picker_flutter: ^1.6.3
   flutter_chat_bubble: ^2.0.2
@@ -30,7 +30,7 @@ dependencies:
   url_launcher:
   pull_to_refresh: ^2.0.0
   flutter_flavor: ^3.1.3
-  flutter_local_notifications: ^16.1.0
+  flutter_local_notifications: ^16.2.0
   rxdart: ^0.27.7
   provider: ^6.1.1
   freezed: ^2.4.5
@@ -46,13 +46,18 @@ dependencies:
   localization: ^2.1.1
   webview_flutter: ^3.0.4 #DONT UPDATE THIS
 
+  connectivity_plus: ^5.0.2
+  hive: ^2.2.3
+  hive_flutter: ^1.1.0
+
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
   flutter_lints: ^3.0.1
-  sealed_generators:
   freezed_annotation: ^2.4.1
+  freezed: ^2.4.5
+  hive_generator: ^2.0.1
   build_runner:
 
 flutter:
