@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+import 'package:hive/hive.dart';
 import 'package:makula_oem/helper/graphQL/graph_ql_config.dart';
 import 'package:makula_oem/helper/model/get_current_user_details_model.dart';
 import 'package:makula_oem/helper/model/get_status_response.dart';
@@ -9,6 +10,7 @@ import 'package:makula_oem/helper/utils/app_preferences.dart';
 import 'package:makula_oem/helper/utils/colors.dart';
 import 'package:makula_oem/helper/utils/constants.dart';
 import 'package:makula_oem/helper/utils/extension_functions.dart';
+import 'package:makula_oem/helper/utils/offline_resources.dart';
 import 'package:makula_oem/helper/utils/routes.dart';
 import 'package:makula_oem/helper/utils/utils.dart';
 import 'package:makula_oem/helper/viewmodels/login_view_model.dart';
@@ -18,12 +20,15 @@ import 'package:makula_oem/views/widgets/makula_edit_text.dart';
 import 'package:makula_oem/views/widgets/makula_text_view.dart';
 import 'package:provider/provider.dart';
 
-final TextEditingController _emailController = TextEditingController();
-final TextEditingController _passwordController = TextEditingController();
+import '../../../../helper/utils/hive_resources.dart';
+
+final TextEditingController _emailController = TextEditingController(text: "azfar.rashid@mmmtechltd.com");
+final TextEditingController _passwordController = TextEditingController(text: "RyCJxRdy0u");
 final FocusNode _emailFieldFocus = FocusNode();
 final FocusNode _passwordFieldFocus = FocusNode();
 late BuildContext _context;
 final appPreferences = AppPreferences();
+
 
 class LoginScreen extends StatefulWidget {
   const LoginScreen({super.key});
@@ -248,10 +253,11 @@ void _signIn() async {
 
 _saveUserToken(LoginMobile data) async {
   console("_saveUserToken");
-  appPreferences.setString(AppPreferences.TOKEN, data.token.toString());
-  appPreferences.setString(
-      AppPreferences.REFRESH_TOKEN, data.refreshToken.toString());
-  appPreferences.setBool(AppPreferences.LOGGED_IN, true);
+  HiveResources.loginBox?.put(OfflineResources.LOGIN_TOKEN_RESPONSE, data);
+  //appPreferences.setString(AppPreferences.TOKEN, data.token.toString());
+ // appPreferences.setString(
+   //   AppPreferences.REFRESH_TOKEN, data.refreshToken.toString());
+  //appPreferences.setBool(AppPreferences.LOGGED_IN, true);
   GraphQLConfig.token = data.token.toString();
   GraphQLConfig.refreshToken = data.refreshToken.toString();
   _getCurrentUserDetails();
@@ -274,8 +280,8 @@ _getCurrentUserDetails() async {
 }
 
 _saveUserDetailsInAppPreferences(CurrentUser data) async {
-  console("_getCurrentUserDetails => ${data.email}");
-  await appPreferences.setData(AppPreferences.USER, data);
+  HiveResources.currentUserBox?.put(OfflineResources.CURRENT_USER_RESPONSE, data);
+  //await appPreferences.setData(AppPreferences.USER, data);
   _getOEMStatues();
   // if (_context.mounted) {
   //   Navigator.of(_context).pushNamedAndRemoveUntil(
@@ -306,7 +312,8 @@ _getOEMStatues() async {
 
 _saveOEMStatues(StatusData response) async {
   console("_saveOEMStatues => ${response.listOwnOemOpenTickets?.length}");
-  await appPreferences.setData(AppPreferences.STATUES, response);
+  HiveResources.oemStatusBox?.put(OfflineResources.OEM_STATUS_RESPONSE, response);
+  //await appPreferences.setData(AppPreferences.STATUES, response);
   if (_context.mounted) {
     Navigator.of(_context).pushNamedAndRemoveUntil(
         dashboardScreenRoute, (Route<dynamic> route) => false);
